Create modularized TeleportDispenser Ignition and Gazebo plugins
Modify Gazebo plugin to mimic asynchronous control flow of the Ignition
plugin. Refactor both plugins to separate the ROS node and message
passing from the simulation aspects, moving the former to the new
TeleportDispenserCommon class.

Currently, the TeleportDispenser state and any publishers/subscribers
are in the TeleportDispenserCommonClass, while any functions to locate
and move objects are located in the respective Ignition/Gazebo plugin
classes.